requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.3
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
pdfplumber>=0.10.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
flask-socketio>=5.3.6
simple-websocket>=1.0.0
gevent>=23.9.0
gevent-websocket>=0.10.1
